+++ custom_components/printer_control_center/studio_profiles.py
@@ -16,7 +16,7 @@ from homeassistant.helpers.storage import Store
 STORAGE_VERSION = 1
 STORAGE_KEY = "printer_control_center.studio_profiles"
 
-BANK_VERSION = "5.0.0-beta7"
+BANK_VERSION = "5.0.0-beta8"
 
 
 DEFAULT_PROFILE_BANK: dict[str, Any] = {